refactor(blorktools): util repack
