+++ packages/blorktools/src/asset_debugger/debugger-scene/debugger-scene.js
@@ -5,7 +5,7 @@ import { initAssetPanel } from '../panels/asset-panel/asset-panel.js';
 import { initHtmlEditorModal } from '../modals/html-editor-modal/html-editor-modal.js';
 import { initWorldPanel } from '../panels/world-panel/world-panel.js';
 import { getState, printStateReport, hasFiles } from '../util/state/scene-state.js';
-import { initUiManager } from '../util/ui/ui-manager.js';
+import { initUiManager } from '../util/scene/ui-manager.js';
 import { hideLoadingSplash, showLoadingSplash, updateLoadingProgress } from '../loading-splash/loading-splash.js';
 import { terminateAllWorkers } from '../util/workers/worker-manager.js';
 import { CSS3DDebugController } from '../util/scene/css3d-debug-controller.js';
@@ -15,7 +15,7 @@ import * as rigController from '../util/scene/rig/rig-controller.js';
 import * as lightingManager from '../util/scene/lighting-manager.js';
 import * as backgroundManager from '../util/scene/background-manager.js';
 import * as modelHandler from '../util/scene/threejs-model-manager.js';
-import * as textureHandler from '../util/upload/handlers/texture-file-handler.js';
+import * as textureHandler from '../util/data/upload/handlers/texture-file-handler.js';
 import * as worldPanel from '../panels/world-panel/world-panel.js';
 import * as assetPanel from '../panels/asset-panel/asset-panel.js';
 import * as htmlEditorModule from '../modals/html-editor-modal/html-editor-modal.js';
