+++ packages/blorktools/src/asset_debugger/util/animation/render/css3d-prerender-controller.js
@@ -1,5 +1,5 @@
 import { showStatus } from "../../../modals/html-editor-modal/html-editor-modal";
-import { createMeshInfoPanel } from "../../ui/mesh-info-panel";
+import { createMeshInfoPanel } from "../../../panels/mesh-info-panel";
 import { updateMeshTexture } from "../playback/animation-playback-controller";
 import { 
     animationCaptureStartTime,
