+++ packages/blorktools/src/asset_debugger/util/data/upload/handlers/model/glb-file-handler.js
@@ -1,4 +1,4 @@
-import { processModelFile } from "../../../workers/worker-manager";
+import { processModelFile } from "../../../../workers/worker-manager";
 
 /**
  * Process a GLB model file using web workers
