+++ packages/blorktools/src/asset_debugger/util/data/upload/zip-handler.js
@@ -6,8 +6,8 @@
 
 // Import JSZip from CDN instead of npm package
 // We'll dynamically import it when needed
-import { TextureClassifier } from '../data/texture-classifier';
-import { updateState } from '../state/scene-state';
+import { TextureClassifier } from '../../data/texture-classifier';
+import { updateState } from '../../state/scene-state';
 
 // Supported image formats for atlas textures
 const SUPPORTED_IMAGE_FORMATS = ['.png', '.jpg', '.jpeg', '.webp', '.tiff', '.tif'];
